--- services/dolphin/manager/src/main/java/com/mobigen/dolphin/ManagerApplication.java
@@ -1,11 +1,0 @@
-package com.mobigen.dolphin;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class ManagerApplication {
-    public static void main(String[] args) {
-        SpringApplication.run(ManagerApplication.class, args);
-    }
-}
