--- services/dolphin/manager/src/main/java/com/mobigen/dolphin/config/MainDBConfiguration.java
@@ -1,64 +1,0 @@
-package com.mobigen.dolphin.config;
-
-import org.springframework.boot.context.properties.ConfigurationProperties;
-import org.springframework.boot.jdbc.DataSourceBuilder;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.context.annotation.Primary;
-import org.springframework.data.jpa.repository.config.EnableJpaRepositories;
-import org.springframework.orm.jpa.JpaTransactionManager;
-import org.springframework.orm.jpa.LocalContainerEntityManagerFactoryBean;
-import org.springframework.orm.jpa.vendor.HibernateJpaVendorAdapter;
-import org.springframework.transaction.PlatformTransactionManager;
-
-import javax.sql.DataSource;
-import java.util.HashMap;
-import java.util.Map;
-
-/**
- * <p>
- * Created by fwani.
- *
- * @version 0.0.1
- * @since 0.0.1
- */
-@Configuration
-@EnableJpaRepositories(
-        basePackages = "com.mobigen.dolphin.repositories",
-        entityManagerFactoryRef = "entityManager",
-        transactionManagerRef = "transactionManager"
-)
-public class MainDBConfiguration {
-    @Bean
-    @Primary
-    @ConfigurationProperties(prefix = "spring.datasource")
-    public DataSource dataSource() {
-        return DataSourceBuilder.create()
-                .build();
-    }
-
-    @Bean
-    @Primary
-    public LocalContainerEntityManagerFactoryBean entityManager() {
-        var em = new LocalContainerEntityManagerFactoryBean();
-        em.setDataSource(dataSource());
-        em.setPackagesToScan(new String[]{"com.mobigen.dolphin.model", "com.mobigen.dolphin.models"});
-        var vendorAdapter = new HibernateJpaVendorAdapter();
-        em.setJpaVendorAdapter(vendorAdapter);
-
-        //Hibernate 설정
-        Map<String, Object> properties = new HashMap<>();
-        properties.put("hibernate.hbm2ddl.auto", "none");
-        properties.put("hibernate.dialect", "org.hibernate.dialect.H2Dialect");
-        em.setJpaPropertyMap(properties);
-        return em;
-    }
-
-    @Bean
-    @Primary
-    public PlatformTransactionManager transactionManager() {
-        var transactionManager = new JpaTransactionManager();
-        transactionManager.setEntityManagerFactory(entityManager().getObject());
-        return transactionManager;
-    }
-}
