--- services/dolphin/manager/src/main/java/com/mobigen/dolphin/controller/ApiController.java
@@ -1,58 +1,0 @@
-package com.mobigen.dolphin.controller;
-
-import com.mobigen.dolphin.model.request.ExecuteDto;
-import com.mobigen.dolphin.model.response.ModelDto;
-import com.mobigen.dolphin.service.ModelService;
-import com.mobigen.dolphin.service.QueryService;
-import lombok.RequiredArgsConstructor;
-import org.springframework.web.bind.annotation.*;
-
-import java.util.List;
-import java.util.UUID;
-
-/**
- * <p>
- * Created by fwani.
- *
- * @version 0.0.1
- * @since 0.0.1
- */
-@RequiredArgsConstructor
-@RestController
-@RequestMapping("/dolphin/v1")
-public class ApiController {
-    private final QueryService queryService;
-    private final ModelService modelService;
-    private final String aa = UUID.randomUUID().toString();
-
-    @GetMapping("/model")
-    public List<ModelDto> getModels() {
-        return modelService.getModels();
-    }
-
-    @PostMapping("/model")
-    public ModelDto addModel(@RequestBody ModelDto modelDto) {
-        modelService.getConnectorInfo();
-        return null;
-    }
-
-    @PostMapping("/query/execute")
-    public Object execute(@RequestBody ExecuteDto executeDto) {
-        return queryService.execute(executeDto.getQuery());
-    }
-
-    @GetMapping("/query/read/{job_id:^[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$}")
-    public Object read(@PathVariable("job_id") String jobId) {
-        return null;
-    }
-
-    @GetMapping("/query/download/{job_id:^[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$}")
-    public Object download(@PathVariable("job_id") String jobId) {
-        return null;
-    }
-
-    @GetMapping("/query/status/{job_id:^[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$}")
-    public Object status(@PathVariable("job_id") String jobId) {
-        return queryService.status(UUID.fromString(jobId));
-    }
-}
