--- services/dolphin/manager/src/main/java/com/mobigen/dolphin/model/request/ExecuteDto.java
@@ -1,17 +1,0 @@
-package com.mobigen.dolphin.model.request;
-
-import lombok.Getter;
-import lombok.Setter;
-
-/**
- * <p>
- * Created by fwani.
- *
- * @version 0.0.1
- * @since 0.0.1
- */
-@Getter
-@Setter
-public class ExecuteDto {
-    private String query;
-}
