--- services/dolphin/manager/src/main/java/com/mobigen/dolphin/model/response/ErrorDto.java
@@ -1,20 +1,0 @@
-package com.mobigen.dolphin.model.response;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.Setter;
-
-/**
- * <p>
- * Created by fwani.
- *
- * @version 0.0.1
- * @since 0.0.1
- */
-@Getter
-@Setter
-@AllArgsConstructor
-public class ErrorDto {
-    private int code;
-    private String message;
-}
