--- services/dolphin/manager/src/main/resources/application.yml
@@ -1,8 +1,0 @@
-dolphin:
-  model:
-    catalog: internalhive
-    schema: default
-spring:
-  mvc:
-    pathmatch:
-      matching-strategy: ant_path_matcher
