+++ services/dolphin/settings.gradle.kts
@@ -1,12 +1,12 @@
 // == Define the inner structure of this component ==
 rootProject.name = "dolphin" // the component name
 
-dependencyResolutionManagement {
-    repositories {
-        mavenCentral()
-    }
-}
+//dependencyResolutionManagement {
+//    repositories {
+//        mavenCentral()
+//    }
+//}
 
-include("manager")
-include("executor")
-include("utils")
+//include("manager")
+//include("executor")
+//include("utils")
