+++ services/dolphin/src/main/java/com/mobigen/dolphin/DolphinApplication.java
@@ -1,0 +1,11 @@
+package com.mobigen.dolphin;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class DolphinApplication {
+    public static void main(String[] args) {
+        SpringApplication.run(DolphinApplication.class, args);
+    }
+}
