+++ services/dolphin/src/main/java/com/mobigen/dolphin/entity/request/CheckDto.java
@@ -1,0 +1,19 @@
+package com.mobigen.dolphin.entity.request;
+
+import lombok.Getter;
+
+import javax.validation.constraints.Pattern;
+import java.util.UUID;
+
+/**
+ * <p>
+ * Created by fwani.
+ *
+ * @version 0.0.1
+ * @since 0.0.1
+ */
+@Getter
+public class CheckDto {
+    @Pattern(regexp = "^[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$")
+    private UUID jobId;
+}
