+++ services/dolphin/src/main/java/com/mobigen/dolphin/entity/request/ExecuteDto.java
@@ -1,0 +1,17 @@
+package com.mobigen.dolphin.entity.request;
+
+import lombok.Getter;
+import lombok.Setter;
+
+/**
+ * <p>
+ * Created by fwani.
+ *
+ * @version 0.0.1
+ * @since 0.0.1
+ */
+@Getter
+@Setter
+public class ExecuteDto {
+    private String query;
+}
