+++ services/dolphin/src/main/java/com/mobigen/dolphin/entity/response/ModelDto.java
@@ -1,0 +1,21 @@
+package com.mobigen.dolphin.entity.response;
+
+import lombok.Builder;
+import lombok.Getter;
+
+import java.util.UUID;
+
+/**
+ * <p>
+ * Created by fwani.
+ *
+ * @version 0.0.1
+ * @since 0.0.1
+ */
+@Getter
+@Builder
+public class ModelDto {
+    private UUID id;
+    private String name;
+    private String description;
+}
