--- services/dolphin/utils/build.gradle.kts
@@ -1,3 +1,0 @@
-dependencies {
-    implementation("com.fasterxml.jackson.core:jackson-annotations:2.16.1")
-}
