--- services/dolphin/utils/src/main/java/com/mobigen/dolphin/models/MappingEntity.java
@@ -1,37 +1,0 @@
-package com.mobigen.dolphin.models;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-import org.hibernate.annotations.GenericGenerator;
-import org.hibernate.annotations.Type;
-
-import javax.persistence.*;
-import java.util.UUID;
-
-/**
- * <p>
- * Created by fwani.
- *
- * @version 0.0.1
- * @since 0.0.1
- */
-@Data
-@NoArgsConstructor
-@AllArgsConstructor
-@Builder
-@Entity
-@Table(name = "mapping")
-public class MappingEntity {
-    @Id
-    @GeneratedValue(generator = "uuid")
-    @GenericGenerator(name = "uuid", strategy = "uuid2")
-    @Type(type = "uuid-char")
-    private UUID id;
-    @Column(name = "user_id")
-    private String userId;
-    private String name;
-    @Column(name = "hive_table")
-    private String hiveTable;
-}
