+++ services/settings.gradle.kts
@@ -25,5 +25,5 @@ include("core")
 include("dataLayer")
 include("extraction")
 include("jpaSample")
-includeBuild("dolphin")
+include("dolphin")
 //include("springSample")
